FIX: Namespace project with `upgrade-code` tool
Also make minor adjustments so that the generated .upgrade.yml file
records these correctly when we get around to PSR-2 compliance in future.
