+++ .upgrade.yml
@@ -1,0 +1,6 @@
+mappings:
+  DocumentConverterDecorator: SilverStripe\DocumentConverter\DocumentConverter
+  DocumentImportField: SilverStripe\DocumentConverter\DocumentImportField
+  DocumentImportIFrameField_Importer: SilverStripe\DocumentConverter\DocumentImportIFrameFieldImporter
+  DocumentImportInnerField: SilverStripe\DocumentConverter\DocumentImportInnerField
+  DocumentConverterTest: SilverStripe\DocumentConverter\Tests\DocumentConverterTest
