+++ lang/en.yml
@@ -2,7 +2,7 @@ en:
   DocumentConverter:
     PROCESSFAILED: 'Could not process document, please double-check you uploaded a .doc or .docx format.'
     SERVERUNREACHABLE: 'Could not contact document conversion server. Please try again later or contact your system administrator.'
-  DocumentConverterDecorator:
+  SilverStripe\DocumentConverter\DocumentConverter:
     ImportTab: Import
   DocumentImportField:
     ChooseFolder: 'Choose a folder to save this file'
