+++ lang/zh.yml
@@ -2,8 +2,8 @@ zh:
   DocumentConverter:
     PROCESSFAILED: '无法处理文档，请复核您上传的是 .doc 或 .docx 格式。'
     SERVERUNREACHABLE: '无法连到文档转换服务器。请稍后再试或联系系统管理员。'
-  DocumentConverterDecorator:
-    ImportTab: '导入'
+  SilverStripe\DocumentConverter\DocumentConverter:
+    ImportTab: 导入
   DocumentImportField:
     ChooseFolder: 选择文件夹保存此文件
     EachH1: '每个标题1'
