+++ src/DocumentImportInnerField.php
@@ -1,4 +1,45 @@
 <?php
+
+namespace SilverStripe\DocumentConverter;
+
+
+
+
+
+
+
+use DOMDocument;
+use DOMElement;
+use DOMAttr;
+
+use Page;
+
+
+use Tidy;
+use HtmlEditorConfig;
+use DOMXPath;
+
+
+
+
+use SilverStripe\DocumentConverter\DocumentImportIFrameFieldImporter;
+use SilverStripe\Control\HTTPRequest;
+use SilverStripe\Core\Convert;
+use SilverStripe\Control\HTTPResponse;
+use SilverStripe\Assets\Upload;
+use SilverStripe\Assets\File;
+use SilverStripe\ORM\DataObject;
+use SilverStripe\Core\Config\Config;
+use SilverStripe\DocumentConverter\DocumentImportInnerField;
+use SilverStripe\Core\Injector\Injector;
+use SilverStripe\Assets\Folder;
+use SilverStripe\Assets\FileNameFilter;
+use SilverStripe\Assets\Image;
+use SilverStripe\Control\Director;
+use SilverStripe\Versioned\Versioned;
+use SilverStripe\AssetAdmin\Forms\UploadField;
+
+
 /**
  * DocumentImportInnerField is built on top of UploadField to access a document
  * conversion capabilities. The original field is stripped down to allow only
@@ -28,12 +69,12 @@ class DocumentImportInnerField extends UploadField {
 
 	private static $allowed_actions = array('upload');
 
-	private static $importer_class = 'DocumentImportIFrameField_Importer';
+	private static $importer_class = DocumentImportIFrameFieldImporter::class;
 
 	/**
 	 * Process the document immediately upon upload.
 	 */
-	public function upload(SS_HTTPRequest $request) {
+	public function upload(HTTPRequest $request) {
 		if($this->isDisabled() || $this->isReadonly()) return $this->httpError(403);
 
 		// Protect against CSRF on destructive action
@@ -75,7 +116,7 @@ class DocumentImportInnerField extends UploadField {
 			}
 		}
 
-		$response = new SS_HTTPResponse(Convert::raw2json(array($return)));
+		$response = new HTTPResponse(Convert::raw2json(array($return)));
 		$response->addHeader('Content-Type', 'text/plain');
 		return $response;
 	}
@@ -223,7 +264,7 @@ class DocumentImportInnerField extends UploadField {
 		);
 
 		$sourcePage = $this->form->getRecord();
-		$importerClass = Config::inst()->get('DocumentImportInnerField', 'importer_class');
+		$importerClass = Config::inst()->get(DocumentImportInnerField::class, 'importer_class');
 		$importer = Injector::inst()->create($importerClass, $fileDescriptor, $chosenFolderID);
 		$content = $importer->import();
 
@@ -252,7 +293,7 @@ class DocumentImportInnerField extends UploadField {
 		$xpath = new DOMXPath($doc);
 
 		// make sure any images are added as Image records with a relative link to assets
-		$folderName = ($chosenFolderID) ? DataObject::get_by_id('Folder', $chosenFolderID)->Name : '';
+		$folderName = ($chosenFolderID) ? DataObject::get_by_id(Folder::class, $chosenFolderID)->Name : '';
 		$imgs = $xpath->query('//img');
 		for($i = 0; $i < $imgs->length; $i++) {
 			$img = $imgs->item($i);
@@ -338,8 +379,21 @@ class DocumentImportInnerField extends UploadField {
 			$this->unusedChildren[$child->ID] = $child;
 		}
 
-		set_error_handler('DocumentImportInnerField_error_handler');
-		global $DocumentImportInnerfield_error;
+		$DocumentImportInnerfield_error;
+
+		$DocumentImportInnerField_error_handler = function ($errno, $errstr, $errfile, $errline) use ( $DocumentImportInnerfield_error ) {
+			global $DocumentImportInnerfield_error;
+			$DocumentImportInnerfield_error = _t(
+				'DocumentConverter.PROCESSFAILED',
+				'Could not process document, please double-check you uploaded a .doc or .docx format.',
+				'Document Converter process Word documents into HTML.'
+			);
+
+			// Do not cascade the error through other handlers
+			return true;
+		};
+
+		set_error_handler($DocumentImportInnerField_error_handler);
 
 		$subtitle = null;
 		$subdoc = new DOMDocument();
@@ -392,139 +446,4 @@ class DocumentImportInnerField extends UploadField {
 
 		$sourcePage->write();
 	}
-}
-
-/**
- * Utility class hiding the specifics of the document conversion process.
- */
-class DocumentImportIFrameField_Importer {
-
-	/**
-	 * Associative array of:
-	 * - name: the full name of the file including the extension.
-	 * - path: the path to the file on the local filesystem.
-	 * - mimeType
-	 */
-	protected $fileDescriptor;
-
-	protected $chosenFolderID;
-
-	protected static $docvert_username;
-
-	protected static $docvert_password;
-
-	protected static $docvert_url;
-
-	public static function set_docvert_username($username = null)  {
-		self::$docvert_username = $username;
-	}
-
-	public static function get_docvert_username() {
-		return self::$docvert_username;
-	}
-
-	public static function set_docvert_password($password = null) {
-		self::$docvert_password = $password;
-	}
-
-	public static function get_docvert_password() {
-		return self::$docvert_password;
-	}
-
-	public static function set_docvert_url($url = null) {
-		self::$docvert_url = $url;
-	}
-
-	public static function get_docvert_url() {
-		return self::$docvert_url;
-	}
-
-	public function __construct($fileDescriptor, $chosenFolderID = null) {
-		$this->fileDescriptor = $fileDescriptor;
-		$this->chosenFolderID = $chosenFolderID;
-	}
-
-	public function import() {
-		$ch = curl_init();
-
-		// PHP 5.5+ introduced CURLFile which makes the '@/path/to/file' syntax deprecated.
-		if(class_exists('CURLFile')) {
-			$file = new CURLFile(
-				$this->fileDescriptor['path'],
-				$this->fileDescriptor['mimeType'],
-				$this->fileDescriptor['name']
-			);
-		} else {
-			$file = '@' . $this->fileDescriptor['path'];
-		}
-
-		curl_setopt_array($ch, array(
-			CURLOPT_URL => self::get_docvert_url(),
-			CURLOPT_USERPWD => sprintf('%s:%s', self::get_docvert_username(), self::get_docvert_password()),
-			CURLOPT_POST => 1,
-			CURLOPT_POSTFIELDS => array('file' => $file),
-			CURLOPT_CONNECTTIMEOUT => 25,
-			CURLOPT_TIMEOUT => 100,
-		));
-
-		$folderName = ($this->chosenFolderID) ? '/'.DataObject::get_by_id('Folder', $this->chosenFolderID)->Name : '';
-		$outname = tempnam(ASSETS_PATH, 'convert');
-		$outzip = $outname . '.zip';
-
-		$out = fopen($outzip, 'w');
-		curl_setopt($ch, CURLOPT_FILE, $out);
-		$returnValue = curl_exec($ch);
-		$status = curl_getinfo($ch, CURLINFO_HTTP_CODE);
-		curl_close($ch);
-		fclose($out);
-		chmod($outzip, 0777);
-
-		if (!$returnValue || ($status != 200)) {
-			return array('error' => _t(
-				'DocumentConverter.SERVERUNREACHABLE',
-				'Could not contact document conversion server. Please try again later or contact your system administrator.',
-				'Document Converter process Word documents into HTML.'
-			));
-		}
-
-		// extract the converted document into assets
-		// you need php zip, i.e. port install php5-zip
-		$zip = new ZipArchive();
-
-		if($zip->open($outzip)) {
-			$zip->extractTo(ASSETS_PATH .$folderName);
-		}
-
-		// remove temporary files
-		unlink($outname);
-		unlink($outzip);
-
-		if (!file_exists(ASSETS_PATH . $folderName . '/index.html')) {
-			return array('error' =>  _t(
-				'DocumentConverter.PROCESSFAILED',
-				'Could not process document, please double-check you uploaded a .doc or .docx format.',
-				'Document Converter process Word documents into HTML.'
-			));
-		}
-
-		$content = file_get_contents(ASSETS_PATH . $folderName . '/index.html');
-
-		unlink(ASSETS_PATH . $folderName . '/index.html');
-
-		return $content;
-	}
-
-}
-
-static $DocumentImportInnerfield_error = '';
-
-function DocumentImportInnerField_error_handler($errno, $errstr, $errfile, $errline) {
-	global $DocumentImportInnerfield_error;
-	$DocumentImportInnerfield_error = _t(
-		'DocumentConverter.PROCESSFAILED',
-		'Could not process document, please double-check you uploaded a .doc or .docx format.',
-		'Document Converter process Word documents into HTML.'
-	);
-
-	return true;
 }
