+++ tests/DocumentConverterTest.php
@@ -1,5 +1,12 @@
 <?php
 
+namespace SilverStripe\DocumentConverter\Tests;
+
+
+use SilverStripe\Dev\SapphireTest;
+
+
+
 class DocumentConverterTest extends SapphireTest
 {
     protected $usesDatabase = true;
